Overhaul plugin system
[These notes written by Eric Kidd.]  Once upon a time, Mephisto had a
plugin system which worked like this:

  http://jamescrisp.blogspot.com/2007/04/tips-for-developing-mephisto-plugins.html

Sven Fuchs replaced the old plugin system with a more powerful plugin
system based on Engines.  Here are some links he recommended describing
that plugin system:

  http://jamescrisp.org/2008/08/25/migrating-mephisto-plugins-to-drax-08/
  http://groups.google.com/group/MephistoBlog/browse_thread/thread/ec7cc25e5720ec9c/8d8b867c8f020557?lnk=gst#8d8b867c8f020557
  http://groups.google.com/group/MephistoBlog/browse_thread/thread/7fd7d1198ef8345e/c000f0e1ea8c490e?lnk=gst#c000f0e1ea8c490e

This is a collapsed, edited version of a plugins branch by Isaac Kearse.
The new API is based on engines, but it has greater compatibility with
the old plugins API.

Much of the code in lib/mephisto/plugin.rb is taken from an earlier
version of Mephisto, and was deleted by Sven Fuchs in revision
1ad1b56b4a6c9284534b0afbb9d5d87715ae4312.

Open questions for code review:
  What does Admin::PluginsController#destroy do?
  Who can access Admin::PluginsController?  Only multi-site admins?
  Do plugin preferences and multi-site support interact reasonably?

v2 (Eric Kidd):
  Collapsed working branch into single patch
  Changed flash message for PluginsController#destroy
  HTML-escaped tab names
  Fixed indentation in unit test files & removed obsolete comments
  Documented spec/models/plugin_spec.rb as out-of-date