Add updated mephisto_feedback plugin
[Notes from Eric Kidd.]  This is a copy of Rick's original
mephisto_feedback plugin, updated by Isaac Kearse for the new plugins
API.  (I'm setting the author of this commit as Rick, because the code
is almost all his.)

This plugin probably won't be included with the default installation of
Mephisto, because it implements a feature that only interests a subset
of Mephisto users.  But it makes a good test case to see how this new
plugin branch works.